# public/swagger.json
{
	"components": {
...
		"schemas": {
			"VerificationMethod": {
				"properties": {
					"id": {
						"type": "string"
					},
					"controller": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"Ed25519VerificationKey2018"
						],
						"nullable": false
					},
					"publicKeyMultibase": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"controller",
					"type",
					"publicKeyMultibase"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Service": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"LinkedDomains",
							"DIDCommMessaging"
						]
					},
					"serviceEndpoint": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"type",
					"serviceEndpoint"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidDocument": {
				"properties": {
					"@context": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"controller": {
						"type": "string"
					},
					"verificationMethod": {
						"items": {
							"$ref": "#/components/schemas/VerificationMethod"
						},
						"type": "array"
					},
					"authentication": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"assertionMethod": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"keyAgreement": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"capabilityInvocation": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"capabilityDelegation": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"service": {
						"items": {
							"$ref": "#/components/schemas/Service"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"@context",
					"verificationMethod",
					"authentication",
					"assertionMethod",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateErrorJSON": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDidDocumentRegisterPayload": {
				"properties": {
					"publicKeyMultibase": {
						"type": "string"
					}
				},
				"required": [
					"publicKeyMultibase"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IVerificationMethodRegisterPayload": {
				"properties": {
					"verificationMethod": {
						"$ref": "#/components/schemas/VerificationMethod"
					}
				},
				"required": [
					"verificationMethod"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVerificationMethodUpdateBody": {
				"properties": {
					"controller": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"Ed25519VerificationKey2018"
						],
						"nullable": false
					},
					"publicKeyMultibase": {
						"type": "string"
					}
				},
				"required": [
					"controller",
					"type",
					"publicKeyMultibase"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVerificationMethodUpdatePayload": {
				"properties": {
					"verificationMethod": {
						"$ref": "#/components/schemas/IVerificationMethodUpdateBody"
					}
				},
				"required": [
					"verificationMethod"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RelationshipTypeType": {
				"type": "string",
				"enum": [
					"authentication",
					"assertionMethod",
					"keyAgreement",
					"capabilityDelegation",
					"capabilityInvocation"
				]
			},
			"IVerificationRelationship": {
				"properties": {
					"id": {
						"type": "string"
					},
					"controller": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"Ed25519VerificationKey2018"
						],
						"nullable": false
					},
					"relationshipType": {
						"$ref": "#/components/schemas/RelationshipTypeType"
					},
					"publicKeyMultibase": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"controller",
					"type",
					"relationshipType",
					"publicKeyMultibase"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVerificationRelationshipRegisterPayload": {
				"properties": {
					"verificationRelationship": {
						"$ref": "#/components/schemas/IVerificationRelationship"
					}
				},
				"required": [
					"verificationRelationship"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVerificationRelationshipUpdateBody": {
				"properties": {
					"controller": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"Ed25519VerificationKey2018"
						],
						"nullable": false
					},
					"publicKeyMultibase": {
						"type": "string"
					}
				},
				"required": [
					"controller",
					"type",
					"publicKeyMultibase"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVerificationRelationshipUpdatePayload": {
				"properties": {
					"verificationRelationship": {
						"$ref": "#/components/schemas/IVerificationRelationshipUpdateBody"
					}
				},
				"required": [
					"verificationRelationship"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"SignedRequestHeader": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"DigestHeader": {
				"type": "apiKey",
				"name": "Digest",
				"in": "header"
			},
			"ExpiresHeader": {
				"type": "apiKey",
				"name": "Expires",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "hedera-appnet",
		"version": "1.0.0",
		"description": "This is a hedera appnet API application",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Meeco",
			"url": "https://meeco.me",
			"email": "‍support@meeco.me"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/did": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Register a new DID document. User provides public key that is going to be added as a delegate key that allows user to modify created DID document later.",
				"summary": "Register a new DID Document.",
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDidDocumentRegisterPayload"
							}
						}
					}
				}
			}
		},
...
		"/did/{did}/verification-methods": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Register a new verification method to the DID document",
				"summary": "Register a new verification method to the DID document",
				"tags": [
					"Verification Method"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Register verification method payload",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVerificationMethodRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/did/{did}/verification-methods/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Update verification method on a DID document",
				"summary": "Update verification method on a DID document",
				"tags": [
					"Verification Method"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "A percent-escaped Service ID string <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231%23key-1",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Update verification method payload",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVerificationMethodUpdatePayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Revoke",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Remove verification method from the DID document",
				"summary": "Remove verification method from the DID document",
				"tags": [
					"Verification Method"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "A percent-escaped Service ID string <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231%23key-1",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/did/{did}/verification-relationships": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Register a new verification relationship to the DID document",
				"summary": "Register a new verification relationship to the DID document",
				"tags": [
					"Verification Relationship"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Register verification relationship payload",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVerificationRelationshipRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/did/{did}/verification-relationships/{relationshipType}/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Update verification relationship on the DID document",
				"summary": "Update verification relationship on the DID document",
				"tags": [
					"Verification Relationship"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "String to specify which relationship type key belongs to",
						"in": "path",
						"name": "relationshipType",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RelationshipTypeType"
						}
					},
					{
						"description": "A percent-escaped Service ID string <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231%23key-1",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Update verification relationship payload",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVerificationRelationshipUpdatePayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Revoke",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Remove verification relationship from the DID document",
				"summary": "Remove verification relationship from the DID document",
				"tags": [
					"Verification Relationship"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "String to specify which relationship type key belongs to",
						"in": "path",
						"name": "relationshipType",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RelationshipTypeType"
						}
					},
					{
						"description": "A percent-escaped Service ID string <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231%23key-1",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}